extract all networking code to networking.[ch]
